+++ src/main/java/com/example/howgarts/repository/CasaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CasaRepository extends JpaRepository<Casa, Integer> {
+public interface CasaRepository extends JpaRepository<Casa, Long> {
 }
