+++ src/main/java/com/example/howgarts/repository/MascotaRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MascotaRepository extends JpaRepository<Mascota, Integer> {
+public interface MascotaRepository extends JpaRepository<Mascota, Long> {
 }
