+++ src/main/java/com/example/howgarts/repository/ProfesorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfesorRepository extends JpaRepository<Profesor, Integer> {
+public interface ProfesorRepository extends JpaRepository<Profesor, Long> {
 }
